lead-crawl-cblc/PayloadCreatorScheduler/requirements.txt
boto3==1.13.19
botocore==1.16.19
apscheduler==3.6.3
python_dateutil==2.8.1
slackclient==2.7.1
